[fix] Correct diclare header of `get_brithdays`

+++ include/calendar.h
@@ -9,7 +9,7 @@ char* MONTHS[12] = {"Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep
 
 extern Date* get_date(const int16_t y, const int16_t m, const int16_t d);
 extern Date* get_date_by_second(const int32_t second);
-extern Date** get_brithdays(const int num_of_brithdays);
+extern Date** get_brithdays(const size_t num_of_brithdays);
 extern Date* get_match(Date** brithdays, size_t length);
 
 extern void free_brithdays(Date** brithdays, size_t len);
